+++ openhab/mounts/tmp/openhab/conf/automation/jsr223/python/personal/core_presence.py
@@ -18,5 +18,5 @@ def trigger_presence_on_motion(event):
 def check_presence(event):
     presenceManagement = ir.getItem("Core_Presence")
     state = get_presence(presenceManagement)
-    if isinstance(presenceManagement.state, UnDefType) or state != presenceManagement.state.intValue():
+    if isinstance(presenceManagement.state, UnDefType) or state != presenceManagement.state.floatValue():
         events.postUpdate(presenceManagement, state)
